requirements.txt
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.3.0
PyTweening==1.0.3
svgwrite==1.4.3
tween==0.0.3